Teste_Estagio/requirements.txt
selenium==4.21.0
psycopg2-binary==2.9.10
pyvirtualdisplay
